Change JournalEditor to only partially apply booking
Previously, the JournalEditor could load entries in one of two ways:

- With `skip_booking = True`, the journal was parsed and sorted with no further
  processing; neither booking nor plugins were run.

- With `skip_booking = False`, which was used for the normal transaction import
  workflow, both booking and plugins were run.

Both options were problematic.  Without booking, missing units in the journal
are not resolved and cannot be used for matching.  With booking and plugins run,
the transactions and postings no longer correspond one-to-one with the actual
contents in the journal, causing JournalEditor to fail to make modifications in
some cases.

With this change, the `skip_booking` option is eliminated.  Instead, booking and
plugins are always fully run in order to obtain a complete list of errors.

However, the JournalEditor itself always operates on a "partially-booked"
representation: this representation is corresponds one-to-one with the
pre-booking list of entries, except that any missing units are substituted with
the interpolated units when possible.