Change metadata format for securities_with_pending_cash transfer transactions
Previously, a separate `ofx_fitid_transfer` key was used in place of `ofx_fitid`
to indicate the OFX FITID value for the extra transfer transactions generated.

Using a separate key meant that the metadata inconsistency criteria of the
matching module did not prevent a posting with an `ofx_fitid` key to be
incorrectly merged with a posting with an `ofx_fitid_transfer` key --- such
merges are never correct.

This commit changes the ofx module to use:

  ofx_fitid: ">XXXX"

in place of:

  ofx_fitid_transfer: "XXXX"