feat(journal_editor): Add support for journal file containing ignored entries
JournalEditor now accepts an optional path to a second top-level journal file
which is treated specially: booking and validation are not performed, and the
entries are stored in a separate `ignored_entries` property.

The new `all_entries` property contains both normal and ignored entries.

This functionality will be used to support "ignoring" an imported entry, which
causes it to be written to this separate ignored journal file, and not included
in the normal journal.  The data sources still use the entries in the ignored
journal file to avoid re-importing entries, however.