Fuzzy matching for posting weights.
In some cases -- for example, matching a pay stub's deduction for a 401(k)
contribution to a mutual-fund purchase transaction from the retirement
account's statement -- it is desirable to match exact currency amounts with
the higher-precision computed weight of a posting at cost.

This change makes that possible by considering two postings to be a match if
their weight is in the same currency and within `fuzzy_match_amount` in value.
The amount defaults to 0.01.

One could imagine it desirable to specify different weights for different
currencies; that exercise is left for the future.