healthequity: Match fund transactions by price in preference to cost.
This solves two problems:
* Sell transactions have no cost specified, but were being matched incorrectly
  because get_key_from_raw_entry gave a cost of None, but get_key_from_posting
  gave a cost of (MISSING, MISSING). This could be fixed by using None instead
  of (MISSING, MISSING) in get_key_from_posting, but that seems strictly worse
  than using price instead, as the latter is highly likely to actually
  match (only) a correct journal entry.

* Buy transactions may be unbalanced as imported by default due to
  rounding (on occasion I've seen HealthEquity round differently than
  beancount wants to, i.e. truncating rather than rounding to the nearest
  thousandth-share on buys). If the user fixes them by manually editing the
  journal to change per-share cost into total cost, they will no longer match
  the input properly. Using the price for matching instead fixes this: it's OK
  for the price not to exactly equal the per-share cost, so there's no reason
  for the user to modify it.

Because we now match investment transactions to existing entries by price
rather than cost, we now also specify price (in addition to cost) for buy
transactions, so that we can continue to match buys. There seems to be no
downside to doing this; the fact that we bought a security at a given price on
a given day strongly suggests that this was a correct price for it on that
day. When matching to the journal, we will fall back to a journal entry's cost
if no price is available, so that transactions imported before this commit
still match.