ofx: properly handle negative-zero amounts on outbound transfers
Code in ofx.py includes an assert to check that the amount on an outbound
transfer posting is negative. However, sometimes OFX files include (apparently
useless) transfers of zero shares of a security. These trigger the assert,
killing beancount_import, even if the zero was properly a negative zero.

This commit changes the assert in question to handle negative zero amounts
correctly.