[amazon] fix pretax adjustments not in shipments
Judging by included testdata invoices, it seems that Amazon invoices
used to include pretax adjustments (e.g. S&H, Subscribe & Save savings,
etc) both in each individual shipment summary and in the final payments
block. But modern Amazon invoices only include these once in the
payments block, not in the shipments info.

Before this diff, the modern-style invoices would result in a
transaction with an error about grand total amount mismatch, because the
pretax adjustments in the payments block were not included in the
calculation of the expected grand total.

This commit fixes that without breaking parsing of the older style of
invoices (all existing tests pass without changes). The logic is that we
track the total amount of payment-block pretax adjustments. If we find
adjustments within shipments, that indicates an older style invoice, and
we assert that those total the same amount as the pretax adjustments in
the payments block, then we subtract the total pretax adjustments for
shipments from the tracked total from payments. For a well-formed
old-style invoice, this should leave us with zero in
`payments_pretax_adjustment_total`, so there ends up being no change in
the expected grand total.