[matching] allow grouping postings of differing sign
My use case for allowing this is importing transactions from Amazon,
which can include e.g. Subscribe & Save or other discounts along with
the cost of the item and tax, all in the same Expenses:FIXME:A account,
balancing a credit card posting for the net cost. In order for this to
match the corresponding transaction from my credit card account, we must
allow grouping the item cost, tax, and discounts into a single
MatchablePosting whose weight is the net cost, which can match the
unknown-expense leg of the credit card transaction. Currently such a
grouped MatchablePosting is disallowed by the "same sign" rule.

I'm curious if there are use cases in which the "same sign" rule is
critical to avoid wrong matches? It seems like the net effect of a set
of postings all to the same account is something that should in general
be considered for matching, even if the postings are not all of the same
sign.

Feedback / counter-examples welcome!