Schwab: add WIRE and SECURITY_TRANSFER
These transfers almost always require rewriting to capture the original
lots. Due to that, we need a way to match the original transaction. The
best way would be to match the original CSV line entirely, but that's
not the design of the importer. Instead, we now have an override for the
amount that you can add (on one posting) to dedupe the original CSV
entry on.