frontend: Display the specific posting account that is uncleared
I initially misunderstood how clearing worked and thought it applied to
a whole transaction. I had an account that I did not think any source
considered authoritative so I unsuccessfully tried to debug why a
different account and posting were not being cleared. If the specific
uncleared account was listed, as in this change, I would have
immediately understood the actual problem.

Also fix type of posting in UnclearedPosting.