ofx: support renumbered accounts
Sometimes an account (e.g. a credit card) changes primary identifiers but
maintains a continuous transaction history and OFX series.  This change makes
it possible to list the old account identifiers for an account in the
`old_account_ids` meta field of an open directive; transactions from OFX files
matching either the old or new account ids will be associated with the account.